+++ CHANGELOG.md
@@ -44,6 +44,8 @@ Changelog
     immediate `float` value. Finally, the `CONVERT` instruction now accepts any valid
     memory reference designator (a `MemoryReference`, a string, or a tuple of type
     `(str, int)`) for both its arguments (@appleby, gh-1010).
+-   Raise an error if a gate with non-constant parameters is provided to `lifted_gate`
+    (@notmgsk, gh-1012).
 
 [v2.11](https://github.com/rigetti/pyquil/compare/v2.10.0...v2.11.0) (September 3, 2019)
 ----------------------------------------------------------------------------------------